+++ templates/Error/error500.php
@@ -10,7 +10,7 @@ use Cake\Core\Configure;
 use Cake\Error\Debugger;
 
 if (Configure::read('debug')) :
-    $this->layout = 'dev_error';
+    $this->setLayout('dev_error');
 
     $this->assign('title', $message);
     $this->assign('templateName', 'error500.php');